Clean up llama_index
