+++ mlflow/llama_index/constant.py
@@ -1,0 +1,1 @@
+FLAVOR_NAME = "llama_index"
